Dashboard Chart addition, User delete fix, Item fixed, SalesController, export addition and sales filter

+++ app/Http/Controllers/UserController.php
@@ -70,6 +70,9 @@ class UserController extends Controller
     public function destroy($id)
     {
         $user = User::find($id);
+        if ($user->id == Auth::user()->id) {
+            return redirect()->route('user.index')->with('error', 'You cannot delete your own account!');
+        }
         $user->delete();
 
         return redirect()->route('user.index');
